Dijkstra codes

+++ src/Dijkstra.java
@@ -4,7 +4,7 @@
 public class Dijkstra {
     // Dijkstra's algorithm to find shortest path from s to all other nodes
     public static int [] dijkstra (WeightedGraph G, int s) {
-        final int [] dist = new int [G.size()];  // shortest known distance from "s"
+        final double [] dist = new double [G.size()];  // shortest known distance from "s"
         final int [] pred = new int [G.size()];  // preceeding node in path
         final boolean [] visited = new boolean [G.size()]; // all false initially
 
@@ -20,7 +20,7 @@ public class Dijkstra {
             final int [] n = G.neighbors (next);
             for (int j=0; j<n.length; j++) {
                 final int v = n[j];
-                final int d = dist[next] + G.getWeight(next,v);
+                final double d = dist[next] + G.getWeight(next,v);
                 if (dist[v] > d) {
                     dist[v] = d;
                     pred[v] = next;
@@ -30,8 +30,8 @@ public class Dijkstra {
         return pred;  // (ignore pred[s]==0!)
     }
 
-    private static int minVertex (int [] dist, boolean [] v) {
-        int x = Integer.MAX_VALUE;
+    private static int minVertex (double [] dist, boolean [] v) {
+        double x = Integer.MAX_VALUE;
         int y = -1;   // graph not connected, or no unvisited vertices
         for (int i=0; i<dist.length; i++) {
             if (!v[i] && dist[i]<x) {y=i; x=dist[i];}
